+++ source/MathTools.h
@@ -39,7 +39,7 @@ namespace gba
     float sin(float a);
     float cos(float a);
     float tan(float a);
-    int floor(float n);
+    int16 floor(float n);
     float length(vec2 v);
     vec2 rotateVectorRad(vec2 v, float a);
     float sqrt(float x);
