+++ source/Raycaster.h
@@ -3,19 +3,19 @@ class Raycaster
 {
 	/**** Members ****/
 	public :
-		int cellSize;
+		uint8 cellSize;
 		int* map;
-		int mapSizeX;
-		int mapSizeY;
-		int xResolution;
-		int yResolution;
+		uint16 mapSizeX;
+		uint16 mapSizeY;
+		uint8 xResolution;
+		uint8 yResolution;
 
 	/**** Functions ****/
 	public :
-		Raycaster(int cellSize, int* map, int mapSizeX, int mapSizeY, float fov, int xResolution, int yResolution);
+		Raycaster(uint8 cellSize, int* map, uint16 mapSizeX, uint16 mapSizeY, float fov, uint8 xResolution, uint8 yResolution);
 		void scanEnv(unsigned volatile short* buffer, const gba::vec2 pos, const float angle, const float fov);
 	private :
-		int checkCellValue(gba::vec2 p);
+		int8 checkCellValue(gba::vec2 p);
 		gba::vec2 findHorizontalIntersect(gba::vec2 pos, float angle);
 		gba::vec2 findVerticalIntersect(gba::vec2 pos, float angle);
 		gba::vec2 drawRaycast(gba::vec2 pos, float angle, bool bDebug = false);
